backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
langchain==0.1.12
langgraph==0.0.26
langchain-google-genai==1.0.1
pandas==2.2.1
python-dotenv==1.0.0
pydantic==2.5.3
plotly==5.19.0
python-multipart==0.0.9
google-generativeai==0.4.1